modulo_chatbot/backend/requirements.txt
# Chatbot Backend Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# LLM for routing/light processing
langchain-groq>=0.0.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
